Add UX to the application (#9)
* Changes to some style elements + enabled DPI scaling

* Update for cross device synching

* Small login style change

* Small cursor changes and added style for home page buttons

* Styling update + refactoring of warnings

* Added resultaten page (not complete yet), fixed some functionalities and added some functionalities

* Added UX elements to the app

* UX and a little load balancing

* Small change

* Quality of life update

* Add profile button (#7)

* Add profile in circle icon

* Add profile item to menu

* Handle the profile button click

* Toast change

+++ components/newPatient.py
@@ -5,6 +5,7 @@ from PyQt5 import uic
 from functools import partial
 import os
 import shutil
+import asyncio
 
 from handlers.requestHandlers import makePatientRequest
 
@@ -16,6 +17,7 @@ class NewPatient(QWidget):
         self.page = page
         self.main = main
         self.connectBtn()
+        self.toast.setHidden(True)
         
         self.dateField.setDate(QDate.currentDate())
         
@@ -30,18 +32,25 @@ class NewPatient(QWidget):
         
     def handleBtn(self, name):
         if name == 'continueBtn':
-            status, res = makePatientRequest(self.app.token_type, self.app.token, self.patientNameField.text(), self.emailField.text(), self.dateField.date().toPyDate())
-            if status == 'Ok':
-                self.page.patient_id = res
-                self.main.patientName.setText(self.patientNameField.text())
-                self.patientNameField.setText('')
-                self.main.thread.start()
-                if os.path.isdir("temp"):
-                    shutil.rmtree("temp")
-                os.mkdir("temp")
-                self.page.stackedWidget.setCurrentIndex(2)
-            else: 
-                print(res)
+            if self.patientNameField.text() != '' and self.emailField.text() != '' and self.dateField.date().toPyDate() != '':
+                status, res = asyncio.run(makePatientRequest(self.app.token_type, self.app.token, self.patientNameField.text(), self.emailField.text(), self.dateField.date().toPyDate()))
+                if status == 'Ok':
+                    self.page.patient_id = res['data']['id']
+                    self.main.patientName.setText(self.patientNameField.text())
+                    self.patientNameField.setText('')
+                    self.main.thread.start()
+                    if os.path.isdir("temp"):
+                        shutil.rmtree("temp")
+                    os.mkdir("temp")
+                    self.page.stackedWidget.setCurrentIndex(2)
+                    self.page.main.timer(res['message'])
+                else: 
+                    self.toast.setText(res)
+                    self.toast.setHidden(False)
+            else:
+                self.toast.setText("Alle velden moeten ingevuld zijn!")
+                self.toast.setStyleSheet("background-color: #bd1321;")
+                self.toast.setHidden(False)
             
         if name == "switchBtn":
             self.patientNameField.setText('')
